Move legacy admin routes to section homes (#901)
