+++ docs/architecture/maintenance-log.md
@@ -16,7 +16,7 @@ Tracks when recurring maintenance processes were last run.
 | Data model doc sync | 2026-02-12 | As needed | As needed | — | docs/architecture/data-model.md vs entities |
 | Navigation audit | 2026-03-22 | 2026-04-22 | Monthly | — | `/nav-audit` — discoverability, backlinks |
 | GDPR audit | — | — | Quarterly | — | Exports, consent, PII logging |
-| Migration squash check | 2026-02-24 | 2026-03-24 | Monthly | — | Check `/Admin/DbVersion` on prod, QA (humans.n.burn.camp), and local dev. Oldest `lastApplied` across all three is the safe squash boundary. |
+| Migration squash check | 2026-02-24 | 2026-03-24 | Monthly | — | Check `/Debug/DbVersion` on prod, QA (humans.n.burn.camp), and local dev. Oldest `lastApplied` across all three is the safe squash boundary. |
 | Architecture ratchet sweep (`arch-sweep`) | 2026-05-16 | 2026-05-30 | Bi-weekly | — | Run the ratchet test suite (`dotnet test --filter "FullyQualifiedName~Architecture.Rules"`), report counts per rule, and trend baseline sizes vs. the previous sweep. Goal: every baseline shrinks. New violations get fixed at the source (or, if intentional, surface in PR review when the rule's test goes red). Scanners + baselines: `tests/Humans.Application.Tests/Architecture/{Ratchet,Rules,Baselines}/`. Source rules: `memory/architecture/*.md` (no-cross-section-ef-joins, no-drops-until-prod-verified, repository-required-for-db-access, no-concurrency-tokens, no-linq-at-db-layer, no-startup-guards, display-sort-in-controllers, no-business-logic-in-controllers) and `docs/architecture/code-review-rules.md` (no controller-injects-DbContext). 2026-05-16 sweep (peterdrier/Humans#595 — Stream E): all baselines stable or shrunk vs 2026-05-04 except `NoDestructiveMigrationOps` (+2; new migrations) and `OnlyAuditLogRepositoryWritesAuditLogEntries` (+1 — real regression, follow-up). Two new rules added the same sweep — `ApplicationServiceEntityReadReturns` (52 initial) and `CrossSectionRepositoryInjection` (1) — those numbers are seeds, not regressions. Field-level invariants (DisplayName / Profile.IsSuspended reads) enforced via `[Obsolete]` + CS0618, not separate scanners. /Admin/* route invariant enforced via memory atom `no-admin-url-section` + code review. Established 2026-05-04 in nobodies-collective/Humans#636. |
 | NuGet full update | 2026-05-17 | 2026-06-17 | Monthly | — | Non-security package updates. 2026-05-17: LOW-risk batch — ASP.NET Core / EF Core / DataProtection / Localization / Mvc.Testing / Identity.EFCore 10.0.7 → 10.0.8 (servicing roll incl. ManagedAuthenticatedEncryptor calc fix); `System.Security.Cryptography.Xml` transitive pin 10.0.7 → 10.0.8; Google.Apis.* (Auth/Admin.Directory/CloudIdentity/Drive/DriveActivity/Groupssettings) → 1.74; `Markdig` 1.1.3 → 1.2.0; `Microsoft.NET.Test.Sdk` 18.4.0 → 18.5.1; `Meziantou.Analyzer` 3.0.54 → 3.0.85; `Microsoft.CodeAnalysis.CSharp` 5.0.0 → 5.3.0 (+ companion pins Common/CSharp.Workspaces/Workspaces.Common/Workspaces.MSBuild). Deferred MEDIUM: `Anthropic` 12.11 → 12.21 (10-minor gap on streaming surface) and `Ical.Net` 5.2.1 → 5.2.2 (deprecates RecurrenceRules/RecurrencePattern — needs migration). |
 | About page package sync | 2026-05-17 | 2026-06-17 | Monthly | — | Update `About.cshtml` package versions after NuGet updates |
