+++ docs/architecture/tech-debt-2026-04-23.md
@@ -11,11 +11,11 @@ Most of the original action plan and many audit findings below are now resolved.
 High-level:
 - **Action plan (9 items):** #1, #2, #3, #4, #5, #6, #7, #8, #9 all done. Only partial leftovers remain as noted below.
 - **Critical section:** §15i doc drift fixed; all three flagged services migrated out of Infrastructure (TeamService → Application/Teams, GoogleWorkspaceSyncService → Application/GoogleIntegration, CalendarService → Application/Calendar); cross-domain nav reads now pragma-wrapped; repository pattern flipped on 4 of 6 flagged repos (ApplicationRepository and ShiftSignupRepository intentionally remain Scoped per documented decisions).
-- **Important section:** jobs no longer inject `HumansDbContext` (all 10 cleaned); 2 of 4 controller DbContext leaks cleaned (AdminController + DevLoginController remain); composer services migrated; HumanRedirectController shim removed; IMemoryCache injection cleaned from jobs (still present in view-components by design).
+- **Important section:** jobs no longer inject `HumansDbContext` (all 10 cleaned); 3 of 4 controller DbContext leaks cleaned (DevLoginController remains); composer services migrated; HumanRedirectController shim removed; IMemoryCache injection cleaned from jobs (still present in view-components by design).
 - **Pre-upstream review follow-ups:** structural blockers fixed; recurring-job concurrency-guard sweep complete (commits 050d595d + b82bcf00 — all 16 recurring jobs now carry `[DisableConcurrentExecution]`); `DuplicateAccountService` sequential-only comment added. Most layering/purity items still real (see inline tags).
 
 **Still genuinely open (bite-sized):**
-- `AdminController.AudienceSegmentation` direct DbContext reads (§2a)
+- `UsersAdminController.Audience` direct DbContext reads (§2a) — resolved by `IAdminDatabaseDiagnosticsService`.
 - ~~`src/Humans.Infrastructure/Jobs/ProcessGoogleSyncOutboxJob.cs` missing `[DisableConcurrentExecution]` attribute~~ **[DONE 2026-04-24, commit 050d595d — also swept all 13 other recurring jobs in commit b82bcf00. Every recurring job now carries `[DisableConcurrentExecution(timeoutInSeconds: 300)]`.]**
 - ~~`docs/sections/Teams.md:92–100` stale line citations~~ **[DONE 2026-04-24 — the SystemTeamType enum and SystemTeamIds constants tables were missing the `BarrioLeads = 6` row; added.]**
 - `BudgetRepository.cs` ResponsibleTeam `.Include` at line 629 — pending nav-strip
@@ -96,7 +96,7 @@ These upstream issues are done in `origin/main` but won't close until the next `
 
 **Controllers with direct `HumansDbContext`** (§2a; already tracked at §15i:638):
 
-- `src/Humans.Web/Controllers/AdminController.cs:26` **[PARTIAL 2026-04-24 — still present at :18/:28; the `AudienceSegmentation` action is the biggest remaining reader; needs a cross-section service method (candidate `IUserService.GetAudienceSegmentAsync`).]**
+- `src/Humans.Web/Controllers/AdminController.cs:26` **[DONE 2026-06-06 — AdminController now owns only the dashboard; audience segmentation moved to UsersAdminController and reads through `IAdminDatabaseDiagnosticsService`.]**
 - `src/Humans.Web/Controllers/ProfileController.cs` (also `IMemoryCache` at :59) **[DONE 2026-04-24 — no DbContext dependency remains; verified via grep.]**
 - `src/Humans.Web/Controllers/GoogleController.cs` (also `IMemoryCache` at :26) **[DONE 2026-04-24 — no DbContext dependency remains; verified via grep.]**
 - `src/Humans.Web/Controllers/DevLoginController.cs:57` (dev-only — low) **[OPEN — still present at :48/:60; dev-only path, low priority.]**
