+++ docs/sections/Calendar.md
@@ -139,7 +139,7 @@ The calendar is intentionally open: no resource-based authorization gates edit/d
 
 - Service lives in `Humans.Application/Services/Calendar/CalendarService.cs` and never imports `Microsoft.EntityFrameworkCore` (enforced by the project's reference graph, design-rules §2b).
 - `ICalendarRepository` (impl in `Humans.Infrastructure/Repositories/Calendar/CalendarRepository.cs`) is the only code path that touches `calendar_events` / `calendar_event_exceptions` via `DbContext`.
-- **Caching decorator** — `CachingCalendarService` (Singleton, in `Humans.Infrastructure/Services/Calendar/`) wraps the keyed Scoped inner `ICalendarService` and owns the `CalendarEventInfo` projection — every non-soft-deleted event row with its `Exceptions` collection embedded, keyed by event id. Load-all warmup uses the normal `ICalendarService.GetAllEventInfosAsync` read method; per-key refresh uses `GetEventInfoAsync`. Window queries (`GetOccurrencesInWindowAsync`) are answered by snapshot-scanning the dict and delegating expansion to `CalendarOccurrenceExpander`. All five mutation paths (`Create`, `Update`, `Delete`, `CancelOccurrence`, `OverrideOccurrence`) flow through the decorator, which delegates to the inner and then reloads the affected event through the same read surface to refresh the single cache entry (or removes it if soft-deleted). **Per-occurrence writes (cancel/override) evict the PARENT event entry** — there is no separate cache row for `CalendarEventException`. Documented at the call site (`CachingCalendarService.InvalidateEventAsync` remarks) and on the projection record (`CalendarEventInfo` `<remarks>`). Surfaced on `/Admin/CacheStats` as `Calendar.Event`. `TrackedCache` owns startup warmup; the decorator is registered as the hosted service.
+- **Caching decorator** — `CachingCalendarService` (Singleton, in `Humans.Infrastructure/Services/Calendar/`) wraps the keyed Scoped inner `ICalendarService` and owns the `CalendarEventInfo` projection — every non-soft-deleted event row with its `Exceptions` collection embedded, keyed by event id. Load-all warmup uses the normal `ICalendarService.GetAllEventInfosAsync` read method; per-key refresh uses `GetEventInfoAsync`. Window queries (`GetOccurrencesInWindowAsync`) are answered by snapshot-scanning the dict and delegating expansion to `CalendarOccurrenceExpander`. All five mutation paths (`Create`, `Update`, `Delete`, `CancelOccurrence`, `OverrideOccurrence`) flow through the decorator, which delegates to the inner and then reloads the affected event through the same read surface to refresh the single cache entry (or removes it if soft-deleted). **Per-occurrence writes (cancel/override) evict the PARENT event entry** — there is no separate cache row for `CalendarEventException`. Documented at the call site (`CachingCalendarService.InvalidateEventAsync` remarks) and on the projection record (`CalendarEventInfo` `<remarks>`). Surfaced on `/Debug/CacheStats` as `Calendar.Event`. `TrackedCache` owns startup warmup; the decorator is registered as the hosted service.
 - **Cross-domain navs** — `CalendarEvent.OwningTeam` is `[Obsolete]`-marked per §6c; EF references it under `#pragma warning disable CS0618` in `CalendarEventConfiguration` solely to declare the FK + cascade. Display stitching routes through `ITeamServiceRead.GetTeamsAsync` (§6b in-memory join). Aggregate-local nav `CalendarEvent.Exceptions` is kept and eagerly loaded by the repository.
 - **Cross-section calls** — public interfaces this section consumes: `ITeamServiceRead` (display names, team picker), `IAuditLogService` (mutation audit).
 - **Architecture test** — `tests/Humans.Application.Tests/Architecture/CalendarArchitectureTests.cs` pins the §15 shape and the decorator invariants (`CachingCalendarService` is sealed, implements `ICalendarService`/`ICalendarServiceRead`, surfaces `ICacheStats`, and keeps the read interface DTO-only).
