+++ docs/sections/Users.md
@@ -131,7 +131,7 @@ Two controllers serve this section:
 | Import jobs (Tickets, MailerLite) | Call `IAccountProvisioningService.FindOrCreateUserByEmailAsync` to materialize contact-only User rows (no `LastLoginAt`). |
 | Authenticated human | Read own User row. Self-declare `NotAttending` for the active event year via `IUserService.DeclareNotAttendingAsync` (and undo via `UndoNotAttendingAsync`). Request account deletion. |
 | HumanAdmin, Board, Admin | Read any User's deletion / login state. Suspension itself is a Profile concern (see [Profiles.md](Profiles.md)). |
-| Admin | Trigger account merge flows (see [Profiles.md](Profiles.md) — `AccountMergeService` lives there). Outside Production only: purge a human via `AdminController.PurgeHuman`. |
+| Admin | Trigger account merge flows (see [Profiles.md](Profiles.md) — `AccountMergeService` lives there). Outside Production only: purge a human via `ProfileController.PurgeHuman`. |
 
 ## Invariants
 
@@ -150,8 +150,8 @@ Two controllers serve this section:
 - Application-layer services in `Humans.Application.Services.Users/` **cannot** inject `HumansDbContext` directly — they go through `IUserRepository` / `IUserEmailRepository`. The Application project's reference graph blocks `Microsoft.EntityFrameworkCore`.
 - Other Application-layer services **cannot** read or write the `users` / Identity tables directly — they go through `IUserService`.
 - Regular humans **cannot** purge any account.
-- An Admin **cannot** purge their own account — `AdminController.PurgeHuman` returns the user to the admin-detail page with an error when `user.Id == currentUser.Id`.
-- Purge **cannot** run in Production — `AdminController.PurgeHuman` returns `NotFound` when `IWebHostEnvironment.IsProduction()`. Account anonymization (the GDPR-deletion path through `AnonymizeExpiredAccountAsync`) runs in every environment via `ProcessAccountDeletionsJob`.
+- An Admin **cannot** purge their own account — `ProfileController.PurgeHuman` returns the user to the admin-detail page with an error when `user.Id == currentUser.Id`.
+- Purge **cannot** run in Production — `ProfileController.PurgeHuman` returns `NotFound` when `IWebHostEnvironment.IsProduction()`. Account anonymization (the GDPR-deletion path through `AnonymizeExpiredAccountAsync`) runs in every environment via `ProcessAccountDeletionsJob`.
 - A merge-tombstoned `User` (`MergedToUserId` non-null) **cannot** sign in — `LockoutEnd` is bumped far-future during fold. Application code outside `IAccountMergeService` **cannot** clear `MergedToUserId` / `MergedAt` or revive a tombstone.
 
 ## Triggers
@@ -164,7 +164,7 @@ Two controllers serve this section:
 - **On admin participation backfill:** `IUserService.BackfillParticipationsAsync` writes records with `Source = AdminBackfill`.
 - **On account-deletion request (user-initiated):** `IAccountDeletionService.RequestDeletionAsync` orchestrates the user-initiated path. Internally calls `IUserService.SetDeletionPendingAsync` to stamp `DeletionRequestedAt` + `DeletionScheduledFor` on the User row, revokes team memberships and governance roles immediately so the user loses access during the 30-day grace period, and sends a confirmation email.
 - **On scheduled deletion expiry:** `ProcessAccountDeletionsJob` calls `IAccountDeletionService.AnonymizeExpiredAccountAsync` per due user. That coordinates: end team memberships (`ITeamService.RevokeAllMembershipsAsync`), end governance role assignments (`IRoleAssignmentService.RevokeAllActiveAsync`), anonymize the profile (`IProfileService.AnonymizeExpiredProfileAsync`), cancel active shift signups (`IShiftSignupService.CancelActiveSignupsForUserAsync`), delete VolunteerEventProfile rows (`IShiftManagementService.DeleteShiftProfilesForUserAsync`), then anonymize identity / remove `UserEmail` rows and `AspNetUserLogins` rows (`IUserService.ApplyExpiredDeletionAnonymizationAsync`). Cross-cutting caches (`IFullProfileInvalidator`, Teams active-teams + member, role-assignment claims, shift authorization) are then invalidated. The orchestrator writes the `AuditAction.AccountAnonymized` audit entry and sends the confirmation email.
-- **On admin purge (non-Production only):** `AdminController.PurgeHuman` removes external logins via `UserManager.RemoveLoginAsync`, then calls `IAccountDeletionService.PurgeAsync`. The orchestrator delegates the actual identity collapse to `IUserService.PurgeOwnDataAsync` (anonymizes the `users` row, drops `UserEmail` rows and `AspNetUserLogins` rows, invalidates the FullProfile cache), then drops the Teams active-teams cache and per-user role-assignment / shift-authorization caches. Identity-only — does not cascade to Profile rows. (Replaces the pre-PR routing through `IOnboardingService.PurgeHumanAsync` and the renamed `IUserService.PurgeAsync`.)
+- **On admin purge (non-Production only):** `ProfileController.PurgeHuman` removes external logins via `UserManager.RemoveLoginAsync`, then calls `IAccountDeletionService.PurgeAsync`. The orchestrator delegates the actual identity collapse to `IUserService.PurgeOwnDataAsync` (anonymizes the `users` row, drops `UserEmail` rows and `AspNetUserLogins` rows, invalidates the FullProfile cache), then drops the Teams active-teams cache and per-user role-assignment / shift-authorization caches. Identity-only — does not cascade to Profile rows. (Replaces the pre-PR routing through `IOnboardingService.PurgeHumanAsync` and the renamed `IUserService.PurgeAsync`.)
 - **On account merge accept:** `IAccountMergeService.AcceptAsync` (Profiles section) calls `IUserService.ReassignLoginsToUserAsync` (re-FKs `AspNetUserLogins` source → target) and `ReassignEventParticipationToUserAsync` (re-FKs `event_participations`), then `AnonymizeForMergeAsync` to tombstone the source row by setting `MergedToUserId` + `MergedAt` and bumping `LockoutEnd` far-future so the source can no longer sign in. The source `User` row stays in place as a redirect — it is NOT deleted — so append-only history written under the source id (audit log, consent records, budget audit log) remains attributable.
 - **On per-user reads of a target after merge:** `IUserService.GetMergedSourceIdsAsync(targetUserId)` returns the set of source ids whose `MergedToUserId` points at the target. Append-only sections (`IAuditLogService`, `IConsentService`, `IBudgetService.ContributeForUserAsync`) union this set with `targetUserId` before querying so source-tombstoned rows surface for the target.
