+++ docs/sections/admin-shell.md
@@ -4,7 +4,7 @@ Frame-only section. Provides the shared admin sidebar, breadcrumb, and dashboard
 
 ## Concepts
 
-- The **Admin Shell** is the persistent layout wrapper rendered for every `/Admin/*` page: top-nav, left sidebar, breadcrumb, and page container.
+- The **Admin Shell** is the persistent layout wrapper rendered for the admin dashboard and section admin pages: top-nav, left sidebar, breadcrumb, and page container.
 - The **Sidebar** is the left navigation panel inside the admin shell. It is divided into named groups; each group contains one or more items. Items and groups are filtered at render time by the current user's roles.
 - The **Breadcrumb** is the per-page path strip rendered inside the admin shell header. Each page sets its own breadcrumb via the shared `AdminShell` layout.
 - The **Dashboard skeleton** is the top-level `/Admin` landing page. It aggregates summary stats from multiple sections (humans in review, open feedback, pending shifts, recent audit events) via service calls.
@@ -15,7 +15,7 @@ This section owns no entities.
 
 ## Routing
 
-The admin shell applies to all routes under `/Admin`. The `AdminLayout.cshtml` layout is selected via `_ViewStart.cshtml` for the `Admin` area. Per-page breadcrumb and page title are set via `ViewData["Title"]` and the `AdminBreadcrumb` view component.
+The `/Admin` route is the shared dashboard. The `AdminLayout.cshtml` layout is selected by `_ViewStart.cshtml` in each admin view folder that uses the shell, including section-owned routes such as `/Debug/*`, `/Profile/*/Admin/*`, and `/Campaigns/Admin/*`. Per-page breadcrumb and page title are set via `ViewData["Title"]` and the `AdminBreadcrumb` view component.
 
 ## Actors & Roles
 
@@ -35,7 +35,7 @@ Sidebar groups: Operations, Members, Money, Governance, Integrations, Agent, Peo
 
 ## Invariants
 
-- The `Admin` top-nav link and the `/Admin` dashboard are gated by `PolicyNames.AnyAdminRole` (12 roles: Admin, Board, HumanAdmin, TeamsAdmin, CampAdmin, TicketAdmin, FeedbackAdmin, FinanceAdmin, StoreAdmin, NoInfoAdmin, VolunteerCoordinator, ConsentCoordinator). Other actions on `AdminController` remain `PolicyNames.AdminOnly`.
+- The `Admin` top-nav link and the `/Admin` dashboard are gated by `PolicyNames.AnyAdminRole` (12 roles: Admin, Board, HumanAdmin, TeamsAdmin, CampAdmin, TicketAdmin, FeedbackAdmin, FinanceAdmin, StoreAdmin, NoInfoAdmin, VolunteerCoordinator, ConsentCoordinator). Concrete admin tools are gated on their section controllers.
 - Sidebar items are filtered per-item by `IAuthorizationService.AuthorizeAsync`; an item the current user cannot access does not appear in the rendered HTML.
 - Sidebar groups whose entire visible-item list is empty do not render.
 - The admin shell adds no new authorization policies; it reuses existing `PolicyNames.*` constants defined in the Auth section.
@@ -43,7 +43,7 @@ Sidebar groups: Operations, Members, Money, Governance, Integrations, Agent, Peo
 
 ## Negative Access Rules
 
-- A user with no admin-shaped role **cannot** reach the `/Admin` dashboard — `[Authorize(Policy = PolicyNames.AnyAdminRole)]` on `AdminController.Index` rejects them before the shell renders. Non-Index actions are individually gated, most by `PolicyNames.AdminOnly`.
+- A user with no admin-shaped role **cannot** reach the `/Admin` dashboard: `[Authorize(Policy = PolicyNames.AnyAdminRole)]` on `AdminController.Index` rejects them before the shell renders. Section admin actions are individually gated, most by `PolicyNames.AdminOnly`.
 - An admin-role user **cannot** see sidebar items they are not authorized for — items are individually gated, not globally shown.
 
 ## Triggers
