+++ src/Humans.Application/Interfaces/Caching/ICacheStats.cs
@@ -3,7 +3,7 @@ namespace Humans.Application.Interfaces.Caching;
 /// <summary>
 /// Diagnostic snapshot for a single in-memory caching dictionary owned by a
 /// Singleton decorator (e.g. <c>CachingUserService</c>'s UserInfo dict).
-/// Surfaced on <c>/Admin/CacheStats</c> alongside <see cref="ICacheStatsProvider"/>
+/// Surfaced on <c>/Debug/CacheStats</c> alongside <see cref="ICacheStatsProvider"/>
 /// (which covers <see cref="Microsoft.Extensions.Caching.Memory.IMemoryCache"/>).
 /// </summary>
 public interface ICacheStats
