+++ src/Humans.Application/Interfaces/Caching/TrackedCache.cs
@@ -24,7 +24,7 @@ public class TrackedCache<TKey, TValue> : IHostedService, ICacheStats where TKey
     public string Name { get; }
 
     /// <summary>Constructs a TrackedCache.</summary>
-    /// <param name="name">Stable identifier for /Admin/CacheStats.</param>
+    /// <param name="name">Stable identifier for /Debug/CacheStats.</param>
     /// <param name="warmOnStartup">When true, StartAsync triggers WarmAllAsync; load-all readers call EnsureWarmedAsync on demand.</param>
     /// <param name="logger">Surfaces startup-warmup failures at Warning before they're swallowed — see memory/architecture/no-startup-guards.md.</param>
     public TrackedCache(string name, bool warmOnStartup, ILogger logger)
@@ -44,7 +44,7 @@ public class TrackedCache<TKey, TValue> : IHostedService, ICacheStats where TKey
         ? Math.Round(Hits * 100.0 / (Hits + Misses), 1)
         : 0;
 
-    /// <summary>True once WarmAllAsync has completed; resets on Clear. Readers use EnsureWarmedAsync, not this directly. Surfaced on <c>/Admin/CacheStats</c> as a "Warmed" indicator.</summary>
+    /// <summary>True once WarmAllAsync has completed; resets on Clear. Readers use EnsureWarmedAsync, not this directly. Surfaced on <c>/Debug/CacheStats</c> as a "Warmed" indicator.</summary>
     public bool IsWarmedUp => _warmedUp;
 
     /// <summary>Live snapshot of cached values; concurrent-safe iteration, not a copy.</summary>
