+++ src/Humans.Web/Views/Debug/CacheStats.cshtml
@@ -14,7 +14,7 @@
                 <i class="fa-solid fa-rotate-right me-1"></i>Reset
             </button>
         </form>
-        <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
+        <a asp-controller="Admin" asp-action="Index" class="btn btn-outline-secondary btn-sm">
             <i class="fa-solid fa-arrow-left me-1"></i>Admin
         </a>
     </div>
