+++ src/Humans.Web/Views/Debug/Configuration.cshtml
@@ -5,7 +5,7 @@
 
 <div class="d-flex justify-content-between align-items-center mb-4">
     <h1>Configuration Status</h1>
-    <a asp-action="Index" class="btn btn-outline-secondary">Back to Dashboard</a>
+    <a asp-controller="Admin" asp-action="Index" class="btn btn-outline-secondary">Back to Dashboard</a>
 </div>
 
 @{
