+++ src/Humans.Web/Views/Debug/Logs.cshtml
@@ -30,7 +30,7 @@
 <div class="container-fluid px-4">
     <nav aria-label="breadcrumb" class="mt-4">
         <ol class="breadcrumb">
-            <li class="breadcrumb-item"><a asp-action="Index">Admin</a></li>
+            <li class="breadcrumb-item"><a asp-controller="Admin" asp-action="Index">Admin</a></li>
             <li class="breadcrumb-item active" aria-current="page">Application Logs</li>
         </ol>
     </nav>
