+++ src/Humans.Web/Views/UsersAdmin/_ViewStart.cshtml
@@ -1,0 +1,3 @@
+@{
+    Layout = "_AdminLayout";
+}
