+++ tests/Humans.Application.Tests/Controllers/ProfileControllerDietaryMedicalReplayTests.cs
@@ -111,6 +111,7 @@ public class ProfileControllerDietaryMedicalReplayTests
             Substitute.For<ICampaignService>(),
             Substitute.For<ICampServiceRead>(),
             Substitute.For<IEmailOutboxService>(),
+            Substitute.For<Microsoft.AspNetCore.Hosting.IWebHostEnvironment>(),
             new FakeClock(Instant.FromUtc(2026, 5, 25, 12, 0)),
             authorizationService,
             Substitute.For<IConsentServiceRead>(),
