+++ tests/Humans.Application.Tests/Controllers/ProfileControllerEditTests.cs
@@ -115,6 +115,7 @@ public class ProfileControllerEditTests
             Substitute.For<ICampaignService>(),
             Substitute.For<ICampServiceRead>(),
             Substitute.For<IEmailOutboxService>(),
+            Substitute.For<Microsoft.AspNetCore.Hosting.IWebHostEnvironment>(),
             new FakeClock(Instant.FromUtc(2026, 5, 9, 12, 0)),
             authorizationService,
             Substitute.For<IConsentServiceRead>(),
